test(KlaviyoCore): unify RequestQueue send spy + extract test support
