+++ Tests/KlaviyoCoreTests/RequestQueueTests.swift
@@ -31,297 +31,12 @@ final class RequestQueueTests: XCTestCase {
         super.tearDown()
     }
 
-    // MARK: - Test doubles
-
-    /// A `send` stub that always succeeds. The run-loop tests drive an empty queue, so `flush()`
-    /// never actually calls it, but the actor still requires one.
-    private var alwaysSucceeds: RequestQueue.Send {
-        { _, _ in .success(Data()) }
-    }
-
-    /// A clock that records each requested sleep duration and then suspends the run loop until the
-    /// test consumes one tick. This makes the loop advance one controlled step at a time so the
-    /// stub `flush()` cannot busy-spin thousands of ticks between assertions.
-    private final class GatedSleepClock: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _requested: [TimeInterval] = []
-        // The loop parks on this continuation each tick; the test resumes it to release one step.
-        private var _waiter: CheckedContinuation<Void, Never>?
-
-        var requested: [TimeInterval] {
-            lock.lock(); defer { lock.unlock() }
-            return _requested
-        }
-
-        var clock: SleepClock {
-            SleepClock { [self] seconds in
-                // Resume the parked continuation on cancellation too, so a `stop()` that cancels the
-                // loop mid-sleep never leaks it (which would trip CheckedContinuation's misuse trap).
-                await withTaskCancellationHandler {
-                    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-                        lock.lock()
-                        _requested.append(seconds)
-                        // If cancellation already fired, resume immediately rather than parking a
-                        // continuation nobody will ever release.
-                        if Task.isCancelled {
-                            lock.unlock()
-                            continuation.resume()
-                        } else {
-                            _waiter = continuation
-                            lock.unlock()
-                        }
-                    }
-                } onCancel: {
-                    resumeWaiter()
-                }
-            }
-        }
-
-        /// Releases exactly one parked sleep, letting the loop run one `flush()` and re-park.
-        func releaseOneTick() { resumeWaiter() }
-
-        private func resumeWaiter() {
-            lock.lock()
-            let waiter = _waiter
-            _waiter = nil
-            lock.unlock()
-            waiter?.resume()
-        }
-
-        /// Blocks (bounded) until at least `count` sleeps have been requested.
-        func waitForRequested(atLeast count: Int, timeout: TimeInterval = 2.0) -> Bool {
-            let deadline = Date().addingTimeInterval(timeout)
-            while Date() < deadline {
-                if requested.count >= count { return true }
-                Thread.sleep(forTimeInterval: 0.01)
-            }
-            return requested.count >= count
-        }
-    }
-
-    /// Records every write so `stop()`'s lease-restore behavior is observable via the shared store:
-    /// a `prepend` under the hood calls `save`, so a non-empty `savedBatches` after `stop()` proves a
-    /// restore happened (and its absence proves it did not).
-    private final class WriteSpyDiskIO {
-        private let lock = NSLock()
-        private var _stored: [KlaviyoRequest] = []
-        private var _savedBatches: [[KlaviyoRequest]] = []
-
-        var stored: [KlaviyoRequest] { lock.lock(); defer { lock.unlock() }; return _stored }
-        var savedBatches: [[KlaviyoRequest]] {
-            lock.lock(); defer { lock.unlock() }; return _savedBatches
-        }
-
-        func makeIO() -> QueueStore.DiskIO {
-            QueueStore.DiskIO(
-                load: { [weak self] in self?.stored ?? [] },
-                save: { [weak self] requests in
-                    guard let self else { return }
-                    self.lock.lock(); defer { self.lock.unlock() }
-                    self._stored = requests
-                    self._savedBatches.append(requests)
-                }
-            )
-        }
-    }
-
-    private func makeStore(diskIO: WriteSpyDiskIO = WriteSpyDiskIO()) -> QueueStore {
-        QueueStore(
-            diskIO: diskIO.makeIO(),
-            scheduler: QueueStore.PersistScheduler { _, work in work() },
-            emitWarning: { _ in }
-        )
-    }
-
-    /// Records the id of every request handed to `send`, in call order, so FIFO ordering and
-    /// call-count assertions are observable. Defaults to always-success.
-    private final class SendSpy: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _sentIds: [String] = []
-
-        var sentIds: [String] { lock.lock(); defer { lock.unlock() }; return _sentIds }
-
-        var send: RequestQueue.Send {
-            { [self] request, _ in
-                lock.lock()
-                _sentIds.append(request.id)
-                lock.unlock()
-                return .success(Data())
-            }
-        }
-    }
-
-    /// A `send` stub whose FIRST invocation parks on a continuation until the test releases it, so a
-    /// flush can be held mid-send (with a batch leased in `requestsInFlight`) while a concurrent
-    /// flush is driven against the actor. Subsequent invocations succeed immediately. `started`
-    /// fulfills once the first send has parked, giving the test a bounded signal to synchronize on.
-    private final class ParkingSendSpy: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _sentIds: [String] = []
-        private var _resume: CheckedContinuation<Void, Never>?
-        private var _parked = false
-        let started: XCTestExpectation
-
-        init(started: XCTestExpectation) {
-            self.started = started
-        }
-
-        var sentIds: [String] { lock.lock(); defer { lock.unlock() }; return _sentIds }
-
-        var send: RequestQueue.Send {
-            { [self] request, _ in
-                let shouldPark: Bool = {
-                    lock.lock(); defer { lock.unlock() }
-                    _sentIds.append(request.id)
-                    let firstCall = !_parked
-                    _parked = true
-                    return firstCall
-                }()
-                if shouldPark {
-                    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-                        lock.lock()
-                        _resume = continuation
-                        lock.unlock()
-                        started.fulfill()
-                    }
-                }
-                return .success(Data())
-            }
-        }
-
-        /// Releases the parked first send so flush A can finish draining its batch.
-        func release() {
-            lock.lock()
-            let continuation = _resume
-            _resume = nil
-            lock.unlock()
-            continuation?.resume()
-        }
-    }
-
-    /// Like `ParkingSendSpy`, but its FIRST (parked) invocation resumes with a caller-supplied
-    /// result instead of always succeeding. Used to hold a flush mid-send, run `stop()` across the
-    /// suspension, then resume returning a `.failure` — exercising the join-point guard on the
-    /// failure branches. Subsequent invocations succeed immediately.
-    private final class FailingParkingSendSpy: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _sentIds: [String] = []
-        private var _resume: CheckedContinuation<Void, Never>?
-        private var _parked = false
-        private let result: Result<Data, KlaviyoAPIError>
-        let started: XCTestExpectation
-
-        init(started: XCTestExpectation, result: Result<Data, KlaviyoAPIError>) {
-            self.started = started
-            self.result = result
-        }
-
-        var sentIds: [String] { lock.lock(); defer { lock.unlock() }; return _sentIds }
-
-        var send: RequestQueue.Send {
-            { [self] request, _ in
-                let shouldPark: Bool = {
-                    lock.lock(); defer { lock.unlock() }
-                    _sentIds.append(request.id)
-                    let firstCall = !_parked
-                    _parked = true
-                    return firstCall
-                }()
-                if shouldPark {
-                    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-                        lock.lock()
-                        _resume = continuation
-                        lock.unlock()
-                        started.fulfill()
-                    }
-                    return result
-                }
-                return .success(Data())
-            }
-        }
-
-        /// Releases the parked first send so the flush resumes with the scripted result.
-        func release() {
-            lock.lock()
-            let continuation = _resume
-            _resume = nil
-            lock.unlock()
-            continuation?.resume()
-        }
-    }
-
-    /// A `send` stub driven by a scripted queue of results, consumed in order (the last result
-    /// repeats once the script is exhausted). Records the id of every request sent so retry/dequeue
-    /// behavior is observable. Thread-safe for use from the actor under test.
-    private final class ScriptedSendSpy: @unchecked Sendable {
-        private let lock = NSLock()
-        private var _results: [Result<Data, KlaviyoAPIError>]
-        private var _sentIds: [String] = []
-        private var _sentAttempts: [Int] = []
-
-        init(results: [Result<Data, KlaviyoAPIError>]) {
-            _results = results
-        }
-
-        var sentIds: [String] { lock.lock(); defer { lock.unlock() }; return _sentIds }
-        var sentAttempts: [Int] { lock.lock(); defer { lock.unlock() }; return _sentAttempts }
-
-        var send: RequestQueue.Send {
-            { [self] request, info in
-                lock.lock()
-                _sentIds.append(request.id)
-                _sentAttempts.append(info.attemptNumber)
-                let result = _results.count > 1
-                    ? _results.removeFirst()
-                    : (_results.first ?? .success(Data()))
-                lock.unlock()
-                return result
-            }
-        }
-    }
-
-    /// Builds a `.registerPushToken` request whose payload carries the given token/enablement/
-    /// background, matching the fields `flush()` writes back to `IdentityStore` on success.
-    private func makeRegisterPushTokenRequest(
-        id: String = UUID().uuidString,
-        token: String,
-        enablement: PushEnablement = .authorized,
-        background: PushBackground = .available
-    ) -> KlaviyoRequest {
-        let payload = PushTokenPayload(
-            pushToken: token,
-            enablement: enablement.rawValue,
-            background: background.rawValue,
-            profile: ProfilePayload(anonymousId: "anon-1")
-        )
-        return KlaviyoRequest(id: id, endpoint: .registerPushToken("test-api-key", payload))
-    }
-
-    private func makeCreateProfileRequest(id: String = UUID().uuidString) -> KlaviyoRequest {
-        KlaviyoRequest(
-            id: id,
-            endpoint: .createProfile("test-api-key", CreateProfilePayload(data: .test))
-        )
-    }
-
-    /// A request whose endpoint has `maxRetries == 1`, so a single retry increment (count → 2)
-    /// exceeds the limit — letting the maxRetries-drop path be exercised in one flush.
-    private func makeLowRetryRequest(id: String = UUID().uuidString) -> KlaviyoRequest {
-        KlaviyoRequest(
-            id: id,
-            endpoint: .resolveDestinationURL(
-                trackingLink: URL(string: "https://klaviyo.com")!,
-                profileInfo: ProfilePayload(anonymousId: "anon-1")
-            )
-        )
-    }
-
     // MARK: - flush()
 
     /// `willDrain` runs before the drain, so a request it enqueues is part of the leased batch and
     /// gets sent in this same flush.
     func testWillDrainRunsBeforeDrain() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let spy = SendSpy()
         let willDrainRequest = makeCreateProfileRequest(id: "from-will-drain")
         let queue = RequestQueue(
@@ -338,7 +53,7 @@ final class RequestQueueTests: XCTestCase {
 
     /// Requests are drained and sent head-first in FIFO order.
     func testDrainsAndSendsInFifoOrder() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let spy = SendSpy()
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "one"), persist: .synchronous)
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "two"), persist: .synchronous)
@@ -352,7 +67,7 @@ final class RequestQueueTests: XCTestCase {
 
     /// An empty queue never calls `send`.
     func testEmptyQueueIsNoOp() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let spy = SendSpy()
         let queue = RequestQueue(clock: .immediate, send: spy.send)
 
@@ -364,7 +79,7 @@ final class RequestQueueTests: XCTestCase {
     /// On a successful `.registerPushToken`, the payload's token (and enablement/background) is
     /// written back to the canonical `IdentityStore`.
     func testRegisterSuccessWritesTokenBackToIdentityStore() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let spy = SendSpy()
         QueueStore.shared.enqueue(
             makeRegisterPushTokenRequest(
@@ -386,7 +101,7 @@ final class RequestQueueTests: XCTestCase {
     /// With no apiKey set, `flush()` is a no-op: `send` is never called.
     func testNoApiKeySkipsFlush() async {
         SDKConfigStore.shared.reset() // clear the apiKey seeded in setUp
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let spy = SendSpy()
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "one"), persist: .synchronous)
         let queue = RequestQueue(clock: .immediate, send: spy.send)
@@ -402,9 +117,10 @@ final class RequestQueueTests: XCTestCase {
     /// gated clock, the loop advances one controlled step at a time, so we assert the requested
     /// duration deterministically rather than an unbounded tick count.
     func testStartTicksAtFlushInterval() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.start()
         XCTAssertTrue(gated.waitForRequested(atLeast: 1), "run loop should request its first sleep")
@@ -420,9 +136,10 @@ final class RequestQueueTests: XCTestCase {
 
     /// `start()` then `stop()` completes without hanging or crashing.
     func testStartThenStopCompletes() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.start()
         XCTAssertTrue(gated.waitForRequested(atLeast: 1))
@@ -431,9 +148,10 @@ final class RequestQueueTests: XCTestCase {
 
     /// A `start()` after a `stop()` works — the loop can be restarted.
     func testRestartAfterStop() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.start()
         XCTAssertTrue(gated.waitForRequested(atLeast: 1))
@@ -446,24 +164,26 @@ final class RequestQueueTests: XCTestCase {
 
     /// With an empty in-flight lease, `stop()` must not prepend anything to the shared store.
     func testStopWithEmptyLeaseDoesNotPrepend() async {
-        let spy = WriteSpyDiskIO()
-        QueueStore.register(makeStore(diskIO: spy))
+        let diskSpy = WriteSpyDiskIO()
+        QueueStore.register(makeQueueStore(diskIO: diskSpy))
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.start()
         XCTAssertTrue(gated.waitForRequested(atLeast: 1))
         await queue.stop()
 
-        XCTAssertTrue(spy.savedBatches.isEmpty, "empty lease must not restore anything")
-        XCTAssertTrue(spy.stored.isEmpty, "no requests should have been written")
+        XCTAssertTrue(diskSpy.savedBatches.isEmpty, "empty lease must not restore anything")
+        XCTAssertTrue(diskSpy.stored.isEmpty, "no requests should have been written")
     }
 
     /// `flushNow()` returns without hanging against an empty queue.
     func testFlushNowReturns() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
         await queue.flushNow()
     }
 
@@ -475,9 +195,10 @@ final class RequestQueueTests: XCTestCase {
     /// resumes it finishes its batch, so every originally-queued request is sent exactly once and the
     /// store is left empty — nothing stranded, nothing lost.
     func testConcurrentFlushDoesNotClobberInFlightLease() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let started = expectation(description: "flush A parked mid-send")
-        let spy = ParkingSendSpy(started: started)
+        // ParkingSendSpy equivalent: parkFirstCall=true, default success result.
+        let spy = SendSpy(parkFirstCall: true, started: started)
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "one"), persist: .synchronous)
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "two"), persist: .synchronous)
         let queue = RequestQueue(clock: .immediate, send: spy.send)
@@ -509,8 +230,9 @@ final class RequestQueueTests: XCTestCase {
     /// The NEXT flush tick (simulated by a second `flushNow()`) resends it; on success it dequeues.
     /// The attempt number increments across the two sends, proving retry bookkeeping persisted.
     func testNetworkErrorRetriesOnNextTickAndIncrementsCount() async {
-        QueueStore.register(makeStore())
-        let spy = ScriptedSendSpy(results: [
+        QueueStore.register(makeQueueStore())
+        // ScriptedSendSpy equivalent: scripted results consumed in order.
+        let spy = SendSpy(results: [
             .failure(.networkError(NSError(domain: "test", code: -1))),
             .success(Data())
         ])
@@ -534,9 +256,9 @@ final class RequestQueueTests: XCTestCase {
     /// A rate-limit error sleeps the backoff, then retries the SAME head in place within one flush
     /// (Decision 2: direct sleep). `RecordingSleepClock` returns instantly and records the backoff.
     func testRateLimitSleepsBackoffThenResends() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let recording = RecordingSleepClock()
-        let spy = ScriptedSendSpy(results: [
+        let spy = SendSpy(results: [
             .failure(.rateLimitError(backOff: 8)),
             .success(Data())
         ])
@@ -552,9 +274,9 @@ final class RequestQueueTests: XCTestCase {
 
     /// A server error behaves like a rate-limit: sleep the backoff, then retry the same head in place.
     func testServerErrorSleepsBackoffThenResends() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let recording = RecordingSleepClock()
-        let spy = ScriptedSendSpy(results: [
+        let spy = SendSpy(results: [
             .failure(.serverError(statusCode: 503, backOff: 5)),
             .success(Data())
         ])
@@ -571,8 +293,8 @@ final class RequestQueueTests: XCTestCase {
     /// Once the retry count exceeds `maxRetries` the head is dropped, not resent. The low-retry
     /// endpoint (`maxRetries == 1`) exceeds on the first network error (count → 2).
     func testExceedingMaxRetriesDropsRequest() async {
-        QueueStore.register(makeStore())
-        let spy = ScriptedSendSpy(results: [
+        QueueStore.register(makeQueueStore())
+        let spy = SendSpy(results: [
             .failure(.networkError(NSError(domain: "test", code: -1)))
         ])
         QueueStore.shared.enqueue(makeLowRetryRequest(id: "doomed"), persist: .synchronous)
@@ -589,8 +311,8 @@ final class RequestQueueTests: XCTestCase {
     /// (via `.synchronous` prepend) and the in-memory lease is cleared.
     func testFailureRestoresRemainingLeaseToQueue() async {
         let diskSpy = WriteSpyDiskIO()
-        QueueStore.register(makeStore(diskIO: diskSpy))
-        let spy = ScriptedSendSpy(results: [
+        QueueStore.register(makeQueueStore(diskIO: diskSpy))
+        let spy = SendSpy(results: [
             .failure(.networkError(NSError(domain: "test", code: -1)))
         ])
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "head"), persist: .synchronous)
@@ -611,9 +333,9 @@ final class RequestQueueTests: XCTestCase {
     /// DROP the head (not restore it) without sleeping the backoff, restore the (now-empty) remaining
     /// lease, and return — leaving the store empty and `send` called exactly once.
     func testExceedingMaxRetriesOnBackoffDropsRequest() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let recording = RecordingSleepClock()
-        let spy = ScriptedSendSpy(results: [
+        let spy = SendSpy(results: [
             .failure(.rateLimitError(backOff: 5))
         ])
         QueueStore.shared.enqueue(makeLowRetryRequest(id: "backoff-doomed"), persist: .synchronous)
@@ -634,8 +356,8 @@ final class RequestQueueTests: XCTestCase {
     /// success — leaving the store empty after a single `flushNow()`. This distinguishes the
     /// dequeue+continue path from retry paths that stop+restore the lease.
     func testNonRetryableErrorDequeuesAndContinuesToNextRequest() async {
-        QueueStore.register(makeStore())
-        let spy = ScriptedSendSpy(results: [
+        QueueStore.register(makeQueueStore())
+        let spy = SendSpy(results: [
             .failure(.internalError("non-retryable")),
             .success(Data())
         ])
@@ -655,9 +377,10 @@ final class RequestQueueTests: XCTestCase {
 
     /// `.reachableViaWiFi` sets `flushInterval = 10.0`: the run loop requests 10-second sleeps.
     func testWiFiConnectivitySetsTenSecondInterval() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.networkConnectivityChanged(.reachableViaWiFi)
         XCTAssertTrue(gated.waitForRequested(atLeast: 1), "loop should request its first sleep")
@@ -674,9 +397,10 @@ final class RequestQueueTests: XCTestCase {
 
     /// `.reachableViaWWAN` sets `flushInterval = 30.0`: the run loop requests 30-second sleeps.
     func testWWANConnectivitySetsThirtySecondInterval() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         await queue.networkConnectivityChanged(.reachableViaWWAN)
         XCTAssertTrue(gated.waitForRequested(atLeast: 1), "loop should request its first sleep")
@@ -693,9 +417,10 @@ final class RequestQueueTests: XCTestCase {
 
     /// `.notReachable` cancels the run loop: no new sleeps are recorded after the call.
     func testNotReachableStopsLoop() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         // Start a loop first so we have something to stop.
         await queue.networkConnectivityChanged(.reachableViaWiFi)
@@ -728,11 +453,12 @@ final class RequestQueueTests: XCTestCase {
     /// restore happened as a direct consequence of going offline, not of the flush completing.
     func testNotReachableRestoresInFlightLease() async {
         let diskSpy = WriteSpyDiskIO()
-        QueueStore.register(makeStore(diskIO: diskSpy))
+        QueueStore.register(makeQueueStore(diskIO: diskSpy))
 
         // A parking send spy lets us hold the actor mid-flush so requestsInFlight is populated.
         let started = expectation(description: "flush parked mid-send")
-        let parking = ParkingSendSpy(started: started)
+        // ParkingSendSpy equivalent: parkFirstCall=true, default success result.
+        let parking = SendSpy(parkFirstCall: true, started: started)
 
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "leased"), persist: .synchronous)
 
@@ -764,12 +490,15 @@ final class RequestQueueTests: XCTestCase {
     /// the failure path: no crash, and the request is preserved in the store (restored by stop(),
     /// neither lost nor double-processed).
     func testStopDuringSendWithFailureOutcomePreservesRequest() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let started = expectation(description: "flush parked mid-send")
-        // First send parks; on release it returns a non-retryable failure. If the join-point guard
-        // were missing, the `.dequeue` branch's `removeFirst()` would crash on the empty lease.
-        let parking = FailingParkingSendSpy(started: started,
-                                            result: .failure(.internalError("boom")))
+        // FailingParkingSendSpy equivalent: first call parks and returns .failure(.internalError),
+        // subsequent calls succeed. [X, .success] with "last repeats" gives identical semantics.
+        let parking = SendSpy(
+            results: [.failure(.internalError("boom")), .success(Data())],
+            parkFirstCall: true,
+            started: started
+        )
 
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "leased"), persist: .synchronous)
         let queue = RequestQueue(clock: .immediate, send: parking.send)
@@ -812,7 +541,7 @@ final class RequestQueueTests: XCTestCase {
             SDKConfigStore.shared.update(KlaviyoConfig(apiKey: "test-api-key"))
 
             let diskSpy = WriteSpyDiskIO()
-            QueueStore.register(makeStore(diskIO: diskSpy))
+            QueueStore.register(makeQueueStore(diskIO: diskSpy))
 
             // A gated clock parks the flush inside the backoff sleep. `flushNow()` only enters one
             // clock.sleep — the backoff — so a gated clock parks there deterministically.
@@ -820,7 +549,7 @@ final class RequestQueueTests: XCTestCase {
 
             // Script: first send fails with a rate-limit error (enters backoff), second would succeed
             // but must never be reached because stop() fires while the backoff sleep is parked.
-            let spy = ScriptedSendSpy(results: [
+            let spy = SendSpy(results: [
                 .failure(.rateLimitError(backOff: 60)),
                 .success(Data())
             ])
@@ -863,9 +592,10 @@ final class RequestQueueTests: XCTestCase {
     /// WiFi → WWAN coalesces: the old WiFi loop is cancelled and only the WWAN interval (30 s)
     /// is observed going forward. No two loops race with different intervals.
     func testRestartCoalescesOldLoop() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
         let gated = GatedSleepClock()
-        let queue = RequestQueue(clock: gated.clock, send: alwaysSucceeds)
+        let spy = SendSpy()
+        let queue = RequestQueue(clock: gated.clock, send: spy.send)
 
         // Start on WiFi.
         await queue.networkConnectivityChanged(.reachableViaWiFi)
@@ -892,28 +622,11 @@ final class RequestQueueTests: XCTestCase {
 
     // MARK: - Invalid-field clear
 
-    /// Builds a 422 error-response JSON body with the given source pointer, matching the Klaviyo
-    /// API error envelope format that `classifyFailure` → `parseError` decodes.
-    private func makeInvalidFieldErrorData(pointer: String) -> Data {
-        """
-        {
-            "errors": [{
-                "id": "err-1",
-                "status": 422,
-                "code": "invalid",
-                "title": "Invalid input.",
-                "detail": "Invalid value.",
-                "source": { "pointer": "\(pointer)" }
-            }]
-        }
-        """.data(using: .utf8)!
-    }
-
     /// A 422 with `/data/attributes/email` pointer must nil `IdentityStore.current.email` and
     /// dequeue the request (not restore/resend). Parity: `resetStateAndDequeue` in the reducer.
     /// Seeds externalId + email to verify the read-modify-write clears ONLY the targeted field.
     func testInvalidEmailClearsCanonicalEmailAndDequeues() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
 
         // Seed email AND externalId so we can verify only email is cleared.
         var identity = IdentityStore.shared.current
@@ -923,7 +636,7 @@ final class RequestQueueTests: XCTestCase {
         let seededAnonymousId = IdentityStore.shared.current.anonymousId
 
         let errorData = makeInvalidFieldErrorData(pointer: "/data/attributes/email")
-        let spy = ScriptedSendSpy(results: [
+        let spy = SendSpy(results: [
             .failure(.httpError(422, errorData))
         ])
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "bad-email"), persist: .synchronous)
@@ -947,7 +660,7 @@ final class RequestQueueTests: XCTestCase {
     /// `IdentityStore.current.phoneNumber` and dequeue the request.
     /// Seeds externalId + phoneNumber to verify the read-modify-write clears ONLY the targeted field.
     func testInvalidPhoneClearsCanonicalPhoneAndDequeues() async {
-        QueueStore.register(makeStore())
+        QueueStore.register(makeQueueStore())
 
         // Seed phoneNumber AND externalId so we can verify only phoneNumber is cleared.
         var identity = IdentityStore.shared.current
@@ -957,7 +670,7 @@ final class RequestQueueTests: XCTestCase {
         let seededAnonymousId = IdentityStore.shared.current.anonymousId
 
         let errorData = makeInvalidFieldErrorData(pointer: "/data/attributes/phone_number")
-        let spy = ScriptedSendSpy(results: [
+        let spy = SendSpy(results: [
             .failure(.httpError(422, errorData))
         ])
         QueueStore.shared.enqueue(makeCreateProfileRequest(id: "bad-phone"), persist: .synchronous)
